TaskId теперь i64, все модули используют этот алиас

+++ src/domain/entities.rs
@@ -6,4 +6,4 @@ pub struct Task {
     pub status: bool
 }
 
-pub type TaskId = u32;
+pub type TaskId = i64;
